+++ srslib_framework/src/localization/map/occupancy/OccupancyMapFactory.cpp
@@ -93,9 +93,9 @@ OccupancyMap* OccupancyMapFactory::fromMetadata(const OccupancyMetadata& metadat
 
     auto occupancyIterator = occupancy.begin();
 
-    for (int row = 0; row < metadata_.heightCells; row++)
+    for (size_t row = 0; row < metadata_.heightCells; row++)
     {
-        for (int col = 0; col < metadata_.widthCells; col++)
+        for (size_t col = 0; col < metadata_.widthCells; col++)
         {
             // Convert the 8 bit cost into an integer cost and store it
             int8_t grayLevel = *occupancyIterator;
